+++ src/vfbquery/term_info_queries.py
